+++ Framework-Headers/Frameworks/UIKit._framework/Objc/WebEditingDelegate-Protocol.h
@@ -15,7 +15,7 @@
 - (BOOL)webView:(WebView *)arg1 doCommandBySelector:(SEL)arg2;
 - (BOOL)webView:(WebView *)arg1 shouldApplyStyle:(DOMCSSStyleDeclaration *)arg2 toElementsInDOMRange:(DOMRange *)arg3;
 - (BOOL)webView:(WebView *)arg1 shouldBeginEditingInDOMRange:(DOMRange *)arg2;
-- (BOOL)webView:(WebView *)arg1 shouldChangeSelectedDOMRange:(DOMRange *)arg2 toDOMRange:(DOMRange *)arg3 affinity:(int)arg4 stillSelecting:(BOOL)arg5;
+- (BOOL)webView:(WebView *)arg1 shouldChangeSelectedDOMRange:(DOMRange *)arg2 toDOMRange:(DOMRange *)arg3 affinity:(unsigned long long)arg4 stillSelecting:(BOOL)arg5;
 - (BOOL)webView:(WebView *)arg1 shouldChangeTypingStyle:(DOMCSSStyleDeclaration *)arg2 toStyle:(DOMCSSStyleDeclaration *)arg3;
 - (BOOL)webView:(WebView *)arg1 shouldDeleteDOMRange:(DOMRange *)arg2;
 - (BOOL)webView:(WebView *)arg1 shouldEndEditingInDOMRange:(DOMRange *)arg2;
