+++ Framework-Headers/Frameworks/UIKit._framework/Objc/_UICollectionLayoutSectionGeometryTranslator.h
@@ -10,7 +10,7 @@ __attribute__((visibility("hidden")))
 @interface _UICollectionLayoutSectionGeometryTranslator : NSObject
 {
     struct CGSize _contentSize;
-    int _layoutAxis;
+    unsigned long long _layoutAxis;
     BOOL _layoutRTL;
     struct NSDirectionalEdgeInsets _supplementaryInsets;
     struct NSDirectionalEdgeInsets _sectionInsets;
@@ -28,7 +28,7 @@ __attribute__((visibility("hidden")))
 - (void)_computeFrames;
 - (struct CGRect)contentRectForSectionRect:(struct CGRect)arg1;
 - (id)description;
-- (id)initWithContentSize:(struct CGSize)arg1 layoutAxis:(int)arg2 layoutRTL:(BOOL)arg3 sectionInsets:(struct NSDirectionalEdgeInsets)arg4 supplementaryInsets:(struct NSDirectionalEdgeInsets)arg5;
+- (id)initWithContentSize:(struct CGSize)arg1 layoutAxis:(unsigned long long)arg2 layoutRTL:(BOOL)arg3 sectionInsets:(struct NSDirectionalEdgeInsets)arg4 supplementaryInsets:(struct NSDirectionalEdgeInsets)arg5;
 
 @end
