+++ Framework-Headers/PrivateFrameworks/FitnessUI._framework/Objc/UIScrollView+ScrollTest.h
@@ -17,8 +17,8 @@
 - (void)_incrementForScrollTest;
 - (void)_performScrollTestToBottom:(id)arg1 iterations:(int)arg2 delta:(int)arg3;
 - (void)_performScrollTestToBottom:(id)arg1 iterations:(int)arg2 delta:(int)arg3 length:(int)arg4;
-- (void)_performScrollTestToBottom:(id)arg1 iterations:(int)arg2 delta:(int)arg3 length:(int)arg4 scrollAxis:(int)arg5;
-- (void)_performScrollTestToBottom:(id)arg1 iterations:(int)arg2 delta:(int)arg3 scrollAxis:(int)arg4;
+- (void)_performScrollTestToBottom:(id)arg1 iterations:(int)arg2 delta:(int)arg3 length:(int)arg4 scrollAxis:(unsigned long long)arg5;
+- (void)_performScrollTestToBottom:(id)arg1 iterations:(int)arg2 delta:(int)arg3 scrollAxis:(unsigned long long)arg4;
 - (void)_reenableImplicitAnimationsAfterScrollTest;
 - (void)_suppressImplicitAnimationsForScrollTest;
 @end
